refactored loginform out
logininfo type now also belongs to the context rather than the form
